renamed Frame.js to ChartWrapper.js and also the internal class names
