+++ samples/charts/BarChartConfigSample.js
@@ -16,7 +16,7 @@
  * under the License.
  */
 import React from 'react';
-import { Frame } from './Frame';
+import { Frame } from './ChartWrapper';
 import './css/Table.css';
 import VizG from '../../src/VizG.jsx';
